+++ CHANGELOG.md
@@ -13,7 +13,7 @@ HumHub Changelog
 - Fix #8196: Fix whitespace between notification bar items
 - Fix #8205: Fix permission filter alignment
 - Fix #8197: Don't use the setting "Allow individual topics" for user and space tags
-- Fix #8216: Fix rendering of email as user subtitle
+- Fix #8216: Fix rendering of email or URL profile fields as link in user subtitle
 
 1.18.3 (May 18, 2026)
 ---------------------